feat: small refactor and set basic config values
