+++ src/Layout.jsx
@@ -1,0 +1,12 @@
+import Navbar from "./Navbar";
+
+const Layout = ({ children }) => {
+  return (
+    <>
+      <Navbar />
+      {children}
+    </>
+  );
+};
+
+export default Layout;
