+++ src/Navbar.jsx
@@ -1,0 +1,13 @@
+import "./Navbar.css";
+
+export default function Navbar() {
+  return (
+    <div className="navbar">
+      <nav>
+        <div className="navbar-header">
+          <h1>Alchemy University</h1>
+        </div>
+      </nav>
+    </div>
+  );
+}
